StudentReply + loombok + getters and setters refactoring
