+++ src/main/java/dbp/techcall/booking/models/Booking.java
@@ -4,8 +4,13 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import lombok.*;
 
-
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
 @Table(name="booking", schema = "spring_app")
 public class Booking {
@@ -30,79 +35,4 @@ public class Booking {
 
     @Column(name="link",nullable = true)
     private String link;
-
-
-    //Constructores:
-
-
-    public Booking(Integer id, Integer studentId, Integer courseId, Integer professorId, String date, String status, String link) {
-        this.id = id;
-        this.studentId = studentId;
-        this.courseId = courseId;
-        this.professorId = professorId;
-        this.date = date;
-        this.status = status;
-        this.link = link;
-    }
-
-    //Default constructor
-    public Booking() {
-    }
-
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public Integer getStudentId() {
-        return studentId;
-    }
-
-    public void setStudentId(Integer studentId) {
-        this.studentId = studentId;
-    }
-
-    public Integer getCourseId() {
-        return courseId;
-    }
-
-    public void setCourseId(Integer courseId) {
-        this.courseId = courseId;
-    }
-
-    public Integer getProfessorId() {
-        return professorId;
-    }
-
-    public void setProfessorId(Integer professorId) {
-        this.professorId = professorId;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public void setDate(String date) {
-        this.date = date;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public String getLink() {
-        return link;
-    }
-
-    public void setLink(String link) {
-        this.link = link;
-    }
 }
