+++ src/main/java/dbp/techcall/course/models/Category.java
@@ -1,13 +1,22 @@
 package dbp.techcall.course.models;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import com.fasterxml.jackson.annotation.JsonIdentityInfo;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.ObjectIdGenerators;
+import dbp.techcall.student.student.domain.models.Student;
+import jakarta.persistence.*;
+import lombok.*;
 
+import java.util.Set;
 
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
 @Table(name="category", schema = "spring_app")
+@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
 public class Category {
 
     @Id
@@ -17,32 +26,9 @@ public class Category {
     @Column(name = "title",nullable = false)
     private String title;
 
-    //Constructores:
-    //Constructor
-    public Category(Integer id, String title) {
-        this.id = id;
-        this.title = title;
-    }
+    @ManyToMany(mappedBy = "interests")
+    private Set<Student> students;
 
-    //Default constructor
-    public Category() {
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
 }
