+++ src/main/java/dbp/techcall/course/models/CourseCategory.java
@@ -4,8 +4,14 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import lombok.*;
 
 
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
 @Table(name="course_category", schema = "spring_app")
 public class CourseCategory {
@@ -17,34 +23,4 @@ public class CourseCategory {
     @Id
     @Column(name="category_id", nullable = false)
     private Integer categoryId;
-
-    //Constructor
-    public CourseCategory(Integer courseId, Integer categoryId) {
-        this.courseId = courseId;
-        this.categoryId = categoryId;
-    }
-
-    //Default constructor
-    public CourseCategory() {
-    }
-
-
-    public Integer getCourseId() {
-        return courseId;
-    }
-
-    public void setCourseId(Integer courseId) {
-        this.courseId = courseId;
-    }
-
-    public Integer getCategoryId() {
-        return categoryId;
-    }
-
-    public void setCategoryId(Integer categoryId) {
-        this.categoryId = categoryId;
-    }
-
-
-
 }
