+++ src/main/java/dbp/techcall/professor/post/infrastructure/Conversation.java
@@ -3,7 +3,14 @@ package dbp.techcall.professor.post.infrastructure;
 import dbp.techcall.professor.professor.infrastructure.models.Professor;
 import dbp.techcall.student.student.domain.models.Student;
 import jakarta.persistence.*;
+import lombok.*;
 
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
 @Table(name = "conversation")
 public class Conversation {
@@ -23,48 +30,4 @@ public class Conversation {
     @JoinColumn(name = "professor_id", referencedColumnName = "id", nullable = false)
     private Professor professor;
 
-    // Constructor
-
-    public Conversation() {
-    }
-
-    public Conversation(String status, Student student, Professor professor) {
-        this.status = status;
-        this.student = student;
-        this.professor = professor;
-    }
-
-    // Getters y setters
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public Student getStudent() {
-        return student;
-    }
-
-    public void setStudent(Student student) {
-        this.student = student;
-    }
-
-    public Professor getProfessor() {
-        return professor;
-    }
-
-    public void setProfessor(Professor professor) {
-        this.professor = professor;
-    }
 }
