+++ src/main/java/dbp/techcall/professor/post/infrastructure/ProfessorReply.java
@@ -1,9 +1,16 @@
 package dbp.techcall.professor.post.infrastructure;
 
-import dbp.techcall.professor.professor.infrastructure.Professor;
+import dbp.techcall.professor.professor.infrastructure.models.Professor;
 import jakarta.persistence.*;
+import lombok.*;
+
 import java.time.ZonedDateTime;
 
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
 @Table(name = "professor_reply")
 public class ProfessorReply {
@@ -28,67 +35,4 @@ public class ProfessorReply {
     @ManyToOne
     @JoinColumn(name = "professor_id", referencedColumnName = "id", nullable = false)
     private Professor professor;
-
-    // Constructor
-
-    public ProfessorReply() {
-    }
-
-    public ProfessorReply(ZonedDateTime time, String body, int status, Conversation conversation, Professor professor) {
-        this.time = time;
-        this.body = body;
-        this.status = status;
-        this.conversation = conversation;
-        this.professor = professor;
-    }
-
-    // Getters y setters
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public ZonedDateTime getTime() {
-        return time;
-    }
-
-    public void setTime(ZonedDateTime time) {
-        this.time = time;
-    }
-
-    public String getBody() {
-        return body;
-    }
-
-    public void setBody(String body) {
-        this.body = body;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public void setStatus(int status) {
-        this.status = status;
-    }
-
-    public Conversation getConversation() {
-        return conversation;
-    }
-
-    public void setConversation(Conversation conversation) {
-        this.conversation = conversation;
-    }
-
-    public Professor getProfessor() {
-        return professor;
-    }
-
-    public void setProfessor(Professor professor) {
-        this.professor = professor;
-    }
 }
