+++ src/main/java/dbp/techcall/professor/professor/infrastructure/Review.java
@@ -1,9 +1,17 @@
 package dbp.techcall.professor.professor.infrastructure;
 
-import dbp.techcall.student.student.domain.Student;
+import dbp.techcall.professor.professor.infrastructure.models.Professor;
+import dbp.techcall.student.student.domain.models.Student;
 import jakarta.persistence.*;
+import lombok.*;
+
 import java.util.Objects;
 
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
 @Table(name = "review")
 public class Review {
@@ -11,8 +19,6 @@ public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-
 
     @ManyToOne
     @JoinColumn(name = "student_id", referencedColumnName = "id", nullable = false)
@@ -22,42 +28,6 @@ public class Review {
     @JoinColumn(name = "professor_id", referencedColumnName = "id", nullable = false)
     private Professor professor;
 
-    // Constructor
-
-    public Review() {
-    }
-
-    public Review(Student student, Professor professor) {
-        this.student = student;
-        this.professor = professor;
-    }
-
-    // Getters y setters
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-
-    public Student getStudent() {
-        return student;
-    }
-
-    public void setStudent(Student student) {
-        this.student = student;
-    }
-
-    public Professor getProfessor() {
-        return professor;
-    }
-
-    public void setProfessor(Professor professor) {
-        this.professor = professor;
-    }
 
     @Override
     public boolean equals(Object o) {
