+++ src/main/java/dbp/techcall/professor/professor/infrastructure/School.java
@@ -1,7 +1,14 @@
 package dbp.techcall.professor.professor.infrastructure;
 
 import jakarta.persistence.*;
+import lombok.*;
 
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@RequiredArgsConstructor
 @Entity
 @Table(name = "school")
 public class School {
@@ -16,39 +23,4 @@ public class School {
     @Column(name = "img_url", nullable = false)
     private String imgUrl;
 
-    // Constructor
-
-    public School() {
-    }
-
-    public School(String name, String imgUrl) {
-        this.name = name;
-        this.imgUrl = imgUrl;
-    }
-
-    // Getters y setters
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getImgUrl() {
-        return imgUrl;
-    }
-
-    public void setImgUrl(String imgUrl) {
-        this.imgUrl = imgUrl;
-    }
 }
