+++ src/main/java/dbp/techcall/professor/professor/infrastructure/models/ProfessorCategory.java
@@ -4,12 +4,17 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import lombok.*;
 
-
+@Getter
+@Setter
+@AllArgsConstructor
+@RequiredArgsConstructor
+@NoArgsConstructor
 @Entity
 @Table(name="professor_category", schema = "spring_app")
 public class ProfessorCategory {
-    //Many to many relationship between professor and category
+    //Many-to-many relationship between professor and category
     //Primary key is composed by professorId and categoryId\
     @Id
     @Column(name="professor_id", nullable = false)
@@ -19,36 +24,6 @@ public class ProfessorCategory {
     @Column(name="category_id", nullable = false)
     private Integer categoryId;
 
-    //Constructor
-    public ProfessorCategory(Integer professorId, Integer categoryId) {
-        this.professorId = professorId;
-        this.categoryId = categoryId;
-    }
-
-    //Default constructor
-    public ProfessorCategory() {
-    }
-
-    //getters
-    public Integer getProfessorId() {
-        return professorId;
-    }
-
-    public Integer getCategoryId() {
-        return categoryId;
-    }
-
-    //setters
-    public void setProfessorId(Integer professorId) {
-        this.professorId = professorId;
-    }
-
-    public void setCategoryId(Integer categoryId) {
-        this.categoryId = categoryId;
-    }
-
-
-
     @Override
     public String toString() {
         return "ProfessorCategory{" +
@@ -56,7 +31,5 @@ public class ProfessorCategory {
                 ", categoryId=" + categoryId +
                 '}';
     }
-
-
 
 }
